+++ NEWS.md
@@ -1,8 +1,14 @@
-## XXXX-XX-XX v3.11.0-SNAPSHOT
+## XXXX-XX-XX v4.0.0-SNAPSHOT
+* [MODSOURMAN-1355](https://folio-org.atlassian.net/browse/MODSOURMAN-1355) Replace advisory lock with consistent record ordering to prevent deadlocks in journal record insertions
 * [MODSOURMAN-1304](https://folio-org.atlassian.net/browse/MODSOURMAN-1304) Handle journal records for MARC on Instance errors and COMPLETED INSTANCE/MARC_BIB events
 * [MODSOURMAN-1323](https://folio-org.atlassian.net/browse/MODSOURMAN-1323) Ensure holdings validation to check central ECS tenant
 * [MODDATAIMP-1214](https://folio-org.atlassian.net/browse/MODDATAIMP-1214) Ensure permission that allows update of shared instance and MARC
 * [MODSOURMAN-1136](https://folio-org.atlassian.net/browse/MODSOURMAN-1136) Send Kafka event when JobExecution is cancelled
+* [MODSOURMAN-1353](https://folio-org.atlassian.net/browse/MODSOURMAN-1353) Remove GET /change-manager/parsedRecords endpoint
+* [MODSOURCE-747](https://folio-org.atlassian.net/browse/MODSOURCE-747) Ensure jobExecutionId in headers for import run through SRM API for skipping events using jobExecutionId from headers
+* [MODSOURMAN-1357](https://folio-org.atlassian.net/browse/MODSOURMAN-1357) Fix: Mapping rules migrations were not idempotent
+* [MODSOURMAN-1250](https://folio-org.atlassian.net/browse/MODSOURMAN-1250) Replace dependency on mod-configuration with mod-settings to get tenant timezone
+* [MODSOURMAN-1359](https://folio-org.atlassian.net/browse/MODSOURMAN-1359) Remove duplicate subfield "x" from 455 and 555 fields in the MARC authority mapping rules
 
 ## 2025-03-13 v3.10.0
 * [MODSOURMAN-1246](https://folio-org.atlassian.net/browse/MODSOURMAN-1246) Added data import completion notifications